+++ Project_S/Assets/StaminaBar.cs
@@ -7,12 +7,12 @@ public class StaminaBar : MonoBehaviour
 
     public Slider slider;
 
-    public void Set_Stamina(int stamina) 
+    public void Set_Stamina(float stamina) 
     {
         slider.value = stamina;
     }
 
-    public void Set_Max_Stamina(int stamina)
+    public void Set_Max_Stamina(float stamina)
     {
         slider.maxValue = stamina;
         slider.value = stamina;
